requirements_basic.txt
wandb
controlnet_aux
mediapipe
joblib
pycocotools
chumpy
plyfile
segment_anything
trimesh
pickle5
numpy==1.22.4
clip_retrieval
torch-metrics
smplx
scipy
joblib
omegaconf
lpips
torchmetrics
numba
accelerate
transformers